requirements.txt
beautifulsoup4==4.8.0
django==2.2.5
django-compressor==2.3
django-debug-toolbar==2.0
django-filter==2.2.0
django-money==0.15
django-redis-cache==2.0.0
django-stubs==1.1.0
djangorestframework==3.10.3
djangorestframework-stubs==1.0.0
hiredis==1.0.0
google-cloud-vision==0.39.0
markdown==3.1.1
mypy==0.720
Pillow==6.1.0
psycopg2-binary==2.8.3
pytest-cov==2.7.1
pytest-django==3.5.1
pytest-mypy==0.4.1
pytest-mypy-plugins==1.0.3
redis==3.3.8
requests==2.22.0
xlrd==1.2.0
